Break out definition-related functionality
